--- toydl/core/scalar/function.py
@@ -1,193 +1,0 @@
-from __future__ import annotations
-
-from typing import TYPE_CHECKING, Tuple
-
-if TYPE_CHECKING:
-    from toydl.core.scalar.scalar import Scalar, ScalarLike
-
-import toydl.core.operator as operators
-
-from toydl.core.scalar.context import Context
-from toydl.core.scalar.history import ScalarHistory
-
-
-def wrap_tuple(x):  # type: ignore
-    """Turn a possible value into a tuple"""
-    if isinstance(x, tuple):
-        return x
-    return (x,)
-
-
-class ScalarFunction:
-    """
-    A wrapper for a mathematical function that processes and produces
-    Scalar variables.
-
-    This is a static class and is never instantiated. We use `class`
-    here to group together the `forward` and `backward` code.
-    """
-
-    @classmethod
-    def _backward(cls, ctx: Context, d_out: float) -> Tuple[float, ...]:
-        return wrap_tuple(cls.backward(ctx, d_out))  # type: ignore
-
-    @classmethod
-    def _forward(cls, ctx: Context, *inputs: float) -> float:
-        return cls.forward(ctx, *inputs)  # type: ignore
-
-    @classmethod
-    def apply(cls, *vals: "ScalarLike") -> Scalar:
-        raw_vals = []
-        scalars = []
-        for v in vals:
-            if isinstance(v, Scalar):
-                scalars.append(v)
-                raw_vals.append(v.data)
-            else:
-                scalars.append(Scalar(v))
-                raw_vals.append(v)
-
-        # Create the context.py.
-        ctx = Context(False)
-
-        # Call forward with the variables.
-        c = cls._forward(ctx, *raw_vals)
-        assert isinstance(c, (float, int)), "Expected return type float got %s" % (
-            type(c)
-        )
-
-        # Create a new variable from the result with a new history.
-        back = ScalarHistory(cls, ctx, scalars)  # type: ignore
-        return Scalar(c, back)
-
-
-class Add(ScalarFunction):
-    """Addition function :math:`f(x, y) = x + y`"""
-
-    @staticmethod
-    def forward(ctx: Context, a: float, b: float) -> float:
-        return a + b
-
-    @staticmethod
-    def backward(ctx: Context, d_output: float) -> Tuple[float, ...]:
-        return d_output, d_output
-
-
-class Log(ScalarFunction):
-    """Log function :math:`f(x) = log(x)`"""
-
-    @staticmethod
-    def forward(ctx: Context, a: float) -> float:
-        ctx.save_for_backward(a)
-        return operators.log(a)
-
-    @staticmethod
-    def backward(ctx: Context, d_output: float) -> float:
-        a = ctx.saved_values[0]
-        return operators.log_back(a, d_output)
-
-
-class Mul(ScalarFunction):
-    """Multiplication function"""
-
-    @staticmethod
-    def forward(ctx: Context, a: float, b: float) -> float:
-        ctx.save_for_backward(a, b)
-        return operators.mul(a, b)
-
-    @staticmethod
-    def backward(ctx: Context, d_output: float) -> float:
-        a, b = ctx.saved_values
-        return operators.mul_back(a, b, d_output)
-
-
-class Inv(ScalarFunction):
-    """Inverse function"""
-
-    @staticmethod
-    def forward(ctx: Context, a: float) -> float:
-        ctx.save_for_backward(a)
-        return operators.inv(a)
-
-    @staticmethod
-    def backward(ctx: Context, d_output: float) -> float:
-        a = ctx.saved_values[0]
-        return operators.inv_back(a, d_output)
-
-
-class Neg(ScalarFunction):
-    """Negation function"""
-
-    @staticmethod
-    def forward(ctx: Context, a: float) -> float:
-        return operators.neg(a)
-
-    @staticmethod
-    def backward(ctx: Context, d_output: float) -> float:
-        return operators.neg_back(d_output)
-
-
-class Sigmoid(ScalarFunction):
-    """Sigmoid function"""
-
-    @staticmethod
-    def forward(ctx: Context, a: float) -> float:
-        ctx.save_for_backward(a)
-        return operators.sigmoid(a)
-
-    @staticmethod
-    def backward(ctx: Context, d_output: float) -> float:
-        a = ctx.saved_values[0]
-        return operators.sigmoid_back(a, d_output)
-
-
-class ReLU(ScalarFunction):
-    """ReLU function"""
-
-    @staticmethod
-    def forward(ctx: Context, a: float) -> float:
-        ctx.save_for_backward(a)
-        return operators.relu(a)
-
-    @staticmethod
-    def backward(ctx: Context, d_output: float) -> float:
-        a = ctx.saved_values[0]
-        return operators.relu_back(a, d_output)
-
-
-class Exp(ScalarFunction):
-    """Exp function"""
-
-    @staticmethod
-    def forward(ctx: Context, a: float) -> float:
-        ctx.save_for_backward(a)
-        return operators.exp(a)
-
-    @staticmethod
-    def backward(ctx: Context, d_output: float) -> float:
-        a = ctx.saved_values[0]
-        return operators.exp_back(a, d_output)
-
-
-class LT(ScalarFunction):
-    """Less-than function :math:`f(x) =` 1.0 if x is less than y else 0.0"""
-
-    @staticmethod
-    def forward(ctx: Context, a: float, b: float) -> float:
-        return operators.lt(a, b)
-
-    @staticmethod
-    def backward(ctx: Context, d_output: float) -> Tuple[float, float]:
-        return 0.0, 0.0
-
-
-class EQ(ScalarFunction):
-    """Equal function :math:`f(x) =` 1.0 if x is equal to y else 0.0"""
-
-    @staticmethod
-    def forward(ctx: Context, a: float, b: float) -> float:
-        return operators.eq(a, b)
-
-    @staticmethod
-    def backward(ctx: Context, d_output: float) -> Tuple[float, float]:
-        return 0.0, 0.0
